add project name in cmakelist
